Change ID type to int64

+++ zendesk/ticket.go
@@ -6,7 +6,7 @@ import (
 )
 
 type Ticket struct {
-	Id              *int       `json:"id,omitempty"`
+	Id              *int64     `json:"id,omitempty"`
 	URL             *string    `json:"url,omitempty"`
 	ExternalId      *string    `json:"external_id,omitempty"`
 	Type            *string    `json:"type,omitempty"`
@@ -16,14 +16,14 @@ type Ticket struct {
 	Priority        *string    `json:"priority,omitempty"`
 	Status          *string    `json:"status,omitempty"`
 	Recipient       *string    `json:"recipient,omitempty"`
-	RequesterId     *int       `json:"requester_id,omitempty"`
-	SubmitterId     *int       `json:"submitter_id,omitempty"`
-	AssigneeId      *int       `json:"assignee_id,omitempty"`
-	OrganizationId  *int       `json:"organization_id,omitempty"`
-	GroupId         *int       `json:"group_id,omitempty"`
-	CollaboratorIds *[]int     `json:"collaborator_ids,omitempty"`
-	ForumTopicId    *int       `json:"forum_topic_id,omitempty"`
-	ProblemId       *int       `json:"problem_id,omitempty"`
+	RequesterId     *int64     `json:"requester_id,omitempty"`
+	SubmitterId     *int64     `json:"submitter_id,omitempty"`
+	AssigneeId      *int64     `json:"assignee_id,omitempty"`
+	OrganizationId  *int64     `json:"organization_id,omitempty"`
+	GroupId         *int64     `json:"group_id,omitempty"`
+	CollaboratorIds *[]int64   `json:"collaborator_ids,omitempty"`
+	ForumTopicId    *int64     `json:"forum_topic_id,omitempty"`
+	ProblemId       *int64     `json:"problem_id,omitempty"`
 	HasIncidents    *bool      `json:"has_incidents,omitempty"`
 	DueAt           *time.Time `json:"due_at,omitempty"`
 	CreatedAt       *time.Time `json:"created_at,omitempty"`
@@ -42,7 +42,7 @@ func NewTicketService(client *Client) *TicketService {
 	return &TicketService{client}
 }
 
-func (s *TicketService) Get(id int) (*Ticket, error) {
+func (s *TicketService) Get(id int64) (*Ticket, error) {
 	res := TicketBody{}
 	err := s.client.Get(fmt.Sprintf("/api/v2/tickets/%d.json", id), &res)
 	return res.Ticket, err
