+++ zendesk/user.go
@@ -6,7 +6,7 @@ import (
 )
 
 type User struct {
-	Id                  *int       `json:"id,omitempty"`
+	Id                  *int64     `json:"id,omitempty"`
 	Url                 *string    `json:"url,omitempty"`
 	Name                *string    `json:"name,omitempty"`
 	ExternalId          *string    `json:"external_id,omitempty"`
@@ -18,7 +18,7 @@ type User struct {
 	Shared              *bool      `json:"shared,omitempty"`
 	SharedAgent         *bool      `json:"shared_agent,omitempty"`
 	Locale              *string    `json:"locale,omitempty"`
-	LocaleId            *int       `json:"locale_id,omitempty"`
+	LocaleId            *int64     `json:"locale_id,omitempty"`
 	Time_zone           *string    `json:"time_zone,omitempty"`
 	LastLoginAt         *time.Time `json:"last_login_at,omitempty"`
 	Email               *string    `json:"email,omitempty"`
@@ -26,9 +26,9 @@ type User struct {
 	Signature           *string    `json:"signature,omitempty"`
 	Details             *string    `json:"details,omitempty"`
 	Notes               *string    `json:"notes,omitempty"`
-	OrganizationId      *int       `json:"organization_id,omitempty"`
+	OrganizationId      *int64     `json:"organization_id,omitempty"`
 	Role                *string    `json:"role,omitempty"`
-	CustomerRoleId      *int       `json:"custom_role_id,omitempty"`
+	CustomerRoleId      *int64     `json:"custom_role_id,omitempty"`
 	Moderator           *bool      `json:"moderator,omitempty"`
 	TicketRestriction   *string    `json:"ticket_restriction,omitempty"`
 	OnlyPrivateComments *bool      `json:"only_private_comments,omitempty"`
@@ -49,7 +49,7 @@ func NewUserService(client *Client) *UserService {
 	return &UserService{client}
 }
 
-func (s *UserService) Get(id int) (*User, error) {
+func (s *UserService) Get(id int64) (*User, error) {
 	res := UserBody{}
 
 	err := s.client.Get(fmt.Sprintf("/api/v2/users/%d.json", id), &res)
